work around syntax highlighter stripping leading blank lines
The syntaxhighlighter library strips blank lines at the start of code
block it is trying to highlight. In addition to just being wrong, this
means that line numbers on the source page will not line up with the
real files.

Work around this by providing the library with fake objects that ignore
the attempt to strip these lines. This uses a fake string and a fake
element to hold it. This is an ugly workaround, but allows us to
continue using the syntax highlighter via npm.