xslate: don't try to remove cache dir
On older deployments, the xslate cache would sometimes not correctly
detect new template files, leading to unpredictable errors. Purging of
the cache directory on startup was added to deal with this.

The normal deployment is now done with a docker container, so there is
nothing to clear out at startup. Additionally, the cache directory isn't
being configured to a static path, so it will always be using a newly
created temp directory. Clearing out the template cache is no longer
relevant, so it can be removed.