Upgrade bootstrap to 5.3.5 and remove ttip package
Removing the ttip package is just a start at
removing jquery as a dependency. The browser
can handle titles itself to create popups so I
have just migrated them to use that instead.

We still have a couple of jquery uses in the code:
 - quick command modal popup
 - Dropdown menu for the login and profile button

These are not trivial to remove unlike ttip so
I will not include them in this branch