GH Actions: use fully qualified base tag when updating prod image
The base tag needs to match the one used in the k8s repo exactly, so if
we are using fully qualified names it needs to be fully qualified.

Add an extra step to calculate the full names for each tag as an output,
which can be used both as a condition in the step to update the k8s repo
as well as in its inputs.